requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pyodbc==4.0.39
python-dotenv==1.0.0
redis==4.5.1
